+++ src/driver/mod.rs
@@ -1,9 +1,10 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 pub mod aw9523b;
 pub mod axp2101;
-pub mod ds16b20;
+pub mod ds18b20;
 pub mod ft6336u;
 pub mod ip5306;
+pub mod onewire;
 pub mod pcnt;
 pub mod pps;
 pub mod radio;